libcoap_sys
Function coap_startup_with_feature_checks

pub fn coap_startup_with_feature_checks()

Initialize the CoAP library and additionally perform runtime checks to ensure that required features (as enabled in Cargo.toml) are available.

You should prefer using this function over coap_startup, as without calling this function some of the features enabled using the Cargo features may not actually be available. (NOTE: However, see the below section on safety).

Either this function or coap_startup must be run once before any libcoap function is called.

If you are absolutely 100% certain that all features you require are always available (or are prepared to deal with error return values/different behavior on your own if they aren’t), you may use coap_startup instead.

SAFETY WARNING

In order to preserve backwards compatibility, this method may (for now) skip the feature checks altogether if they aren’t provided by libcoap (which may be the case for libcoap < 4.3.5rc3).

If you want to be absolutely certain that a given feature is available, you must also check the libcoap version to ensure that it is at least 4.3.5rc3.

This behavior will be changed as soon as libcoap 4.3.5 is released, after which libcoap 4.3.5 will become the minimum supported version and these checks will be mandatory.